working changes

+++ NeuralNetwork-master/Neural-Network/seqNeuralNetworkTrainer.cu
@@ -31,26 +31,26 @@ neuralNetworkTrainer::neuralNetworkTrainer( neuralNetwork *nn )	:	NN(nn),
 {
 	//create delta lists
 	//--------------------------------------------------------------------------------------------------------
-	deltaInputHidden = new( double*[NN->nInput + 1] );
+	deltaInputHidden = new( float*[NN->nInput + 1] );
 	for ( int i=0; i <= NN->nInput; i++ ) 
 	{
-		deltaInputHidden[i] = new (double[NN->nHidden]);
+		deltaInputHidden[i] = new (float[NN->nHidden]);
 		for ( int j=0; j < NN->nHidden; j++ ) deltaInputHidden[i][j] = 0;		
 	}
 
-	deltaHiddenOutput = new( double*[NN->nHidden + 1] );
+	deltaHiddenOutput = new( float*[NN->nHidden + 1] );
 	for ( int i=0; i <= NN->nHidden; i++ ) 
 	{
-		deltaHiddenOutput[i] = new (double[NN->nOutput]);			
+		deltaHiddenOutput[i] = new (float[NN->nOutput]);			
 		for ( int j=0; j < NN->nOutput; j++ ) deltaHiddenOutput[i][j] = 0;		
 	}
 
 	//create error gradient storage
 	//--------------------------------------------------------------------------------------------------------
-	hiddenErrorGradients = new( double[NN->nHidden + 1] );
+	hiddenErrorGradients = new( float[NN->nHidden + 1] );
 	for ( int i=0; i <= NN->nHidden; i++ ) hiddenErrorGradients[i] = 0;
 	
-	outputErrorGradients = new( double[NN->nOutput + 1] );
+	outputErrorGradients = new( float[NN->nOutput + 1] );
 	for ( int i=0; i <= NN->nOutput; i++ ) outputErrorGradients[i] = 0;
 }
 
@@ -98,7 +98,7 @@ void neuralNetworkTrainer::enableLogging(const char* filename, int resolution = 
 /*******************************************************************
 * calculate output error gradient
 ********************************************************************/
-inline double neuralNetworkTrainer::getOutputErrorGradient( double desiredValue, double outputValue)
+inline float neuralNetworkTrainer::getOutputErrorGradient( float desiredValue, float outputValue)
 {
 	//return error gradient
 	return outputValue * ( 1 - outputValue ) * ( desiredValue - outputValue );
@@ -107,7 +107,7 @@ inline double neuralNetworkTrainer::getOutputErrorGradient( double desiredValue,
 /*******************************************************************
 * calculate input error gradient
 ********************************************************************/
-double neuralNetworkTrainer::getHiddenErrorGradient( int j )
+float neuralNetworkTrainer::getHiddenErrorGradient( int j )
 {
 	//get sum of hidden->output weights * output error gradients
 	double weightedSum = 0;
@@ -235,7 +235,7 @@ void neuralNetworkTrainer::runTrainingEpoch( vector<dataEntry*> trainingSet )
 /*******************************************************************
 * Propagate errors back through NN and calculate delta values
 ********************************************************************/
-void neuralNetworkTrainer::backpropagate( double* desiredOutputs )
+void neuralNetworkTrainer::backpropagate( float* desiredOutputs )
 {		
 	//modify deltas between hidden and output layers
 	//--------------------------------------------------------------------------------------------------------
